Simple sort on going

+++ disorder/compute_disorder.c
@@ -6,16 +6,16 @@
 /*   By: vinida-s <vinida-s@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/21 17:15:01 by vinicius          #+#    #+#             */
-/*   Updated: 2026/07/11 19:28:11 by vinida-s         ###   ########.fr       */
+/*   Updated: 2026/07/15 00:01:43 by vinida-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-double	compute_disorder(t_stack *a)
+float	compute_disorder(t_stack *a)
 {
-	double	mistakes;
-	double	total_pairs;
+	float	mistakes;
+	float	total_pairs;
 	t_node	*current_node;
 	t_node	*other_node;
